qiskit-docs-mcp-server/pyproject.toml
[project]
name = "qiskit-docs-mcp-server"
version = "0.1.0"
description = "MCP server for Qiskit documentation retrieval and search"
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.8.0",
    "certifi>=2025.1.31",
    "click>=8.1.8",
    "distro>=1.9.0",
    "fastmcp>=2.8.1,<3",
    "h11>=0.14.0",
    "httpcore>=1.0.7",
    "httpx>=0.28.1",
    "idna>=3.10",
    "jiter>=0.8.2",
    "langchain>=1.2.7",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "modelcontextprotocol>=0.4.0",
    "openai>=1.61.1",
    "pydantic>=2.10.6",
    "pydantic-core>=2.27.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.0.1",
    "sniffio>=1.3.1",
    "starlette>=0.45.3",
    "tqdm>=4.67.1",
    "typing-extensions>=4.12.2",
    "uvicorn>=0.34.0",
]

[[project.authors]]
name = "Quantum+AI Team. IBM Quantum"
email = "Quantum.Plus.AI@ibm.com"

[project.license]
text = "Apache-2.0"

[project.urls]
Homepage = "https://github.com/Qiskit/mcp-servers"
Repository = "https://github.com/Qiskit/mcp-servers"
Documentation = "https://github.com/Qiskit/mcp-servers/tree/main/qiskit-docs-mcp-server#readme"
"Bug Tracker" = "https://github.com/Qiskit/mcp-servers/issues"

[project.scripts]
qiskit-docs-mcp-server = "qiskit_docs_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qiskit_docs_mcp_server"]

keywords = ["quantum", "qiskit", "mcp", "ai", "documentation", "rag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[dependency-groups]
dev = [
    "bandit[toml]>=1.8.2",
    "mypy>=1.15.0",
    "pre-commit>=4.1.0",
    "ruff>=0.9.4",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

[project.optional-dependencies]
examples = [
    "fastmcp>=2.14.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.7",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=1.1.0",
    "langchain-ibm>=0.1.0",
    "langchain-mcp-adapters>=0.0.5",
    "langchain-ollama>=0.1.0",
    "langchain-openai>=0.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.bandit]
skips = ["B101"]
exclude_dirs = ["tests", ".venv", "build", "dist"]
